+++ NEWS.md
@@ -1,6 +1,6 @@
 # Rampy News
 
-Copyright (c) 2014-2021 Dr. Charles Le Losq
+Copyright (c) 2014-2023 Dr. Charles Le Losq
 
 email: lelosq@ipgp.fr
 
@@ -12,11 +12,23 @@ Licence: see LICENCE.md
 
 - Classification via 1D CNN and other algorithms using Keras or Pytorch
 
-- Addition of external calibration method in rameau
-
 - peak fitting for maps
 
-# 0.4.9 (stable)
+# 0.5.0 (stable)
+
+- map() function is now maps() to avoid any conflict with built-in Python map() function
+
+- add the possiblity of importing lines of spectra for Horiba spectrometers
+
+- add a despiking() function
+
+- add external calibration in rameau
+
+- add the posibility to calibrate K coefficient(s) on another dataset of spectra
+
+- correction of a typo in mlclassificator function
+
+# 0.4.9
 
 - quick fix of a bug in read_horiba function (from rampy.maps)